requirements.txt
# aiohttp==3.9.5
# aiosignal==1.3.1
# annotated-types==0.7.0
# anyio==4.4.0
# async-timeout==4.0.3
# attrs==23.2.0
# blinker==1.8.2
# certifi==2024.6.2
# charset-normalizer==3.3.2
# click==8.1.7
# dataclasses-json==0.6.7
# distro==1.9.0
# exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==4.0.1
# frozenlist==1.4.1
google-cloud-storage==2.18.2
gunicorn==22.0.0
# h11==0.14.0
# httpcore==1.0.5
# httpx==0.27.0
# idna==3.7
# importlib_metadata==8.0.0
# itsdangerous==2.2.0
# Jinja2==3.1.4
# jsonpatch==1.33
# jsonpointer==3.0.0
Markdown==3.6
# MarkupSafe==2.1.5
# marshmallow==3.21.3
# multidict==6.0.5
# mypy-extensions==1.0.0
# numpy==1.26.4
openai==1.37.1
# orjson==3.10.5
# packaging==24.1
# pydantic==2.7.4
# pydantic_core==2.18.4
python-dotenv==1.0.1
# PyYAML==6.0.1
# regex==2024.7.24
requests==2.32.3
# sniffio==1.3.1
# SQLAlchemy==2.0.31
# tenacity==8.4.2
# tiktoken==0.7.0
# tqdm==4.66.4
# typing-inspect==0.9.0
# typing_extensions==4.12.2
# urllib3==2.2.2
# Werkzeug==3.0.3
# yarl==1.9.4
# zipp==3.19.2
